(Hopefully) Fix spurious wakeups under read timeouts
I received a bug report about a random mouse tracking sequence being
written to the text buffer after clicking on the LF/CRLF button.
Noticeably, there was no "Esc" character visualizer in the buffer.
This leads me to believe that the VT parser ran into the ESC state
timeout handling (to detect lone ESC chars) and we had a spurious wakeup
from the stdin read() afterward, which resulted in a empty return value,
and lead the VT parser to believe it was truly an ESC key press.