Do not render View menu when no documents are active
The status bar is blank when no documents are active, which makes sense
since all the values on it are document specific. There for we can
remove focus status bar from the View menu when no documents are active.

Since this means all options in the View menu would be gone, we can also
remove the menu entirely. Which is what we do for Edit also.