Always reset wants_statusbar_focus
Currently wants_statusbar_focus will only be reset once we actually
steal focus, but this only happens if we're properly drawing the status
bar. So if we're not drawing it, such as when no documents are open, we
would keep this option active, until we finally do draw it, then we will
switch focus unexpectedly.